+++ examples/jigsaw/Parser.php
@@ -1,0 +1,9 @@
+<?php
+
+use sixlive\ParsedownHighlight;
+use Mni\FrontYAML\Markdown\MarkdownParser;
+
+class Parser extends ParsedownHighlight implements MarkdownParser
+{
+    //
+}