+++ CHANGELOG.md
@@ -4,3 +4,5 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Jigsaw example ([#1](https://github.com/sixlive/parsedown-highlight/pull/1))